# contracts/openapi/portal-v1.yaml
openapi: 3.1.0
info:
  title: Enterprise Portal API
  version: 1.0.0
  summary: Contract-first API for the enterprise portal BFF.
  description: |
    Enterprise Portal API v1 is served under `/api/v1` and uses same-origin,
    cookie-backed sessions with CSRF protection for unsafe requests.

    Rate limiting is documented with `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, `X-RateLimit-Reset`, and `Retry-After` response
    headers. Unless otherwise specified, operations may return the standard
    Problem Details responses defined in `components.responses` for 400, 401,
    403, 404, 409, 422, 429, and 500.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/v1/auth/captcha:
    get:
      tags:
        - Auth
      operationId: issueCaptcha
      summary: Issue a login CAPTCHA challenge.
      description: |
        Returns a self-hosted SVG CAPTCHA bound to a one-time Redis challenge.
        Challenges expire according to `CAPTCHA_TTL` and are consumed on login.
      responses:
        "200":
          description: CAPTCHA challenge issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptchaChallenge"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      operationId: login
      summary: Authenticate with username, password, and CAPTCHA.
      description: |
        Validates CAPTCHA before credentials. Creates a server-side session or
        returns an MFA challenge when additional verification is required.
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login completed or MFA challenge issued.
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/logout:
    post:
      tags:
        - Auth
      operationId: logout
      summary: Terminate the current session.
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        "204":
          description: Current session revoked.
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/password/forgot:
    post:
      tags:
        - Auth
      operationId: requestPasswordReset
      summary: Request a password reset.
      description: Always returns an accepted response to avoid account enumeration.
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "202":
          description: Password reset request accepted.
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/password/reset:
    post:
      tags:
        - Auth
      operationId: resetPassword
      summary: Complete a password reset with a reset token.
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "204":
          description: Password reset completed.
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/mfa/verify:
    post:
      tags:
        - Auth
      operationId: verifyLoginMfa
      summary: Verify MFA during login.
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaVerifyRequest"
      responses:
        "200":
          description: MFA verification succeeded and a session was established.
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/mfa/recovery:
    post:
      tags:
        - Auth
      operationId: verifyLoginRecoveryCode
      summary: Verify a recovery code during login.
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaRecoveryRequest"
      responses:
        "200":
          description: Recovery code accepted and a session was established.
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/csrf:
    get:
      tags:
        - Auth
      operationId: getCsrfToken
      summary: Return the current CSRF token.
      description: Issues or refreshes the CSRF token used in `X-XSRF-TOKEN` for unsafe requests.
      security: []
      responses:
        "200":
          description: Current CSRF token.
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CsrfToken"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: PORTALSESSION
      description: Opaque server-side session cookie.
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: CSRF token required for unsafe same-origin requests.
  headers:
    XRateLimitLimit:
      description: Maximum requests allowed in the current rate-limit window.
      schema:
        type: integer
        minimum: 0
    XRateLimitRemaining:
      description: Remaining requests in the current rate-limit window.
      schema:
        type: integer
        minimum: 0
    XRateLimitReset:
      description: UTC epoch seconds when the current rate-limit window resets.
      schema:
        type: integer
        minimum: 0
    RetryAfter:
      description: Seconds to wait before retrying after a rate limit response.
      schema:
        type: integer
        minimum: 0
...
  responses:
    BadRequest:
      description: Malformed request syntax, invalid parameter shape, or unreadable body.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: Missing, expired, or invalid authenticated session.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Forbidden:
      description: Authenticated principal lacks the required permission or CSRF token.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    Conflict:
      description: Request conflicts with current resource state.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    UnprocessableEntity:
      description: Request is syntactically valid but violates validation rules.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    ServiceUnavailable:
      description: Temporary dependency outage (e.g. CAPTCHA store unavailable).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/XRateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/XRateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/XRateLimitReset"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    InternalServerError:
      description: Unexpected server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
  schemas:
    ProblemDetail:
      type: object
      description: RFC 9457 problem details object.
      additionalProperties: true
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          description: Stable application error code.
        traceId:
          type: string
          description: Correlation identifier for diagnostics.
        errors:
          type: array
          description: Field-level validation errors.
          items:
            type: object
            additionalProperties: false
            required:
              - field
              - message
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
...
    UserStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - LOCKED
        - PENDING
    Language:
      type: string
      enum:
        - fa-IR
        - en-US
    Theme:
      type: string
      enum:
        - LIGHT
        - DARK
        - SYSTEM
    Preferences:
      type: object
      additionalProperties: false
      required:
        - language
        - theme
        - timezone
      properties:
        language:
          $ref: "#/components/schemas/Language"
        theme:
          $ref: "#/components/schemas/Theme"
        timezone:
          type: string
          example: Asia/Tehran
        density:
          type: string
          enum:
            - COMFORTABLE
            - COMPACT
          default: COMFORTABLE
    Permission:
      type: object
      additionalProperties: false
      required:
        - code
        - module
        - action
        - description
      properties:
        code:
          type: string
          pattern: "^[A-Z0-9_]+$"
          example: USER_READ
        module:
          type: string
          example: users
        action:
          type: string
          example: read
        description:
          type: string
        system:
          type: boolean
          default: true
    Role:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - name
        - system
        - permissions
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          pattern: "^[A-Z0-9_]+$"
          example: ADMIN
        name:
          type: string
        description:
          type:
            - string
            - "null"
        system:
          type: boolean
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    User:
      type: object
      additionalProperties: false
      required:
        - id
        - username
        - email
        - displayName
        - status
        - preferences
        - roles
        - permissions
        - mfaEnabled
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          minLength: 3
          maxLength: 80
        email:
          type: string
          format: email
        displayName:
          type: string
          minLength: 1
          maxLength: 160
        status:
          $ref: "#/components/schemas/UserStatus"
        preferences:
          $ref: "#/components/schemas/Preferences"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        permissions:
          type: array
          items:
            type: string
            pattern: "^[A-Z0-9_]+$"
        mfaEnabled:
          type: boolean
        lastLoginAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Session:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - createdAt
        - lastAccessedAt
        - expiresAt
        - current
      properties:
        id:
          type: string
        userId:
          type: string
          format: uuid
        username:
          type: string
        ipAddress:
          type: string
        userAgent:
          type: string
        createdAt:
          type: string
          format: date-time
        lastAccessedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        revokedAt:
          type:
            - string
            - "null"
          format: date-time
        current:
          type: boolean
...
    CaptchaChallenge:
      type: object
      additionalProperties: false
      required:
        - captchaId
        - imageSvg
        - expiresAt
        - ttlSeconds
      properties:
        captchaId:
          type: string
          format: uuid
        imageSvg:
          type: string
          description: Inline SVG markup for the CAPTCHA image.
        expiresAt:
          type: string
          format: date-time
        ttlSeconds:
          type: integer
          minimum: 1
        revealAnswer:
          type: string
          description: |
            Present only when `CAPTCHA_REVEAL_ANSWER=true` (automated tests).
            Never enabled in the prod profile.
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - username
        - password
        - captchaId
        - captchaAnswer
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        captchaId:
          type: string
          format: uuid
        captchaAnswer:
          type: string
          minLength: 4
          maxLength: 12
        rememberDevice:
          type: boolean
          default: false
    LoginResponse:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - AUTHENTICATED
            - MFA_REQUIRED
        user:
          $ref: "#/components/schemas/User"
        session:
          $ref: "#/components/schemas/Session"
        mfaChallenge:
          $ref: "#/components/schemas/MfaChallenge"
    MfaChallenge:
      type: object
      additionalProperties: false
      required:
        - challengeId
        - methods
        - expiresAt
      properties:
        challengeId:
          type: string
        methods:
          type: array
          items:
            type: string
            enum:
              - TOTP
              - RECOVERY_CODE
        expiresAt:
          type: string
          format: date-time
    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
        newPassword:
          type: string
          format: password
          minLength: 12
    MfaVerifyRequest:
      type: object
      additionalProperties: false
      required:
        - challengeId
        - code
      properties:
        challengeId:
          type: string
        code:
          type: string
          minLength: 6
          maxLength: 8
        rememberDevice:
          type: boolean
          default: false
    MfaRecoveryRequest:
      type: object
      additionalProperties: false
      required:
        - challengeId
        - recoveryCode
      properties:
        challengeId:
          type: string
        recoveryCode:
          type: string
        rememberDevice:
          type: boolean
          default: false
    CsrfToken:
      type: object
      additionalProperties: false
      required:
        - headerName
        - token
      properties:
        headerName:
          type: string
          const: X-XSRF-TOKEN
        token:
          type: string
...
    ActionResult:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
        id:
          type: string
        completedAt:
          type: string
          format: date-time
...